move functions to base view
