+++ hmpps_person_match_score/views/base_view.py
@@ -1,4 +1,6 @@
 import os
+import uuid
+from typing import List
 
 import flask
 from flask.views import MethodView
@@ -38,3 +40,20 @@ class BaseView(MethodView):
         if not os.path.exists(model_path):
             raise Exception(f"{model_path} does not exist.")
         return model_path
+
+    @staticmethod
+    def generate_view_uuid():
+        return f"v_{uuid.uuid4().hex.replace('-', '')[:16]}"
+
+    def cleanup_splink_tables(self, linker, input_table_alias: List[str]):
+        """
+        Cleans up intermediate tables created by Splink in the DuckDB connection
+        """
+        table_prefixes = ("__splink__df_concat_with_tf_", "__splink__df_predict_")
+
+        for k, v in list(linker._intermediate_table_cache.items()):  # noqa: SLF001
+            if k.startswith(table_prefixes):
+                v.drop_table_from_database_and_remove_from_cache()
+
+        for alias in input_table_alias:
+            self.duckdb_connection.sql(f"DROP VIEW IF EXISTS {alias}")
