+++ hmpps_person_match_score/views/match_view.py
@@ -7,7 +7,6 @@ from splink.duckdb.duckdb_linker import DuckDBLinker
 from hmpps_person_match_score.domain.events import Events
 from hmpps_person_match_score.domain.splink_models import SplinkModels
 from hmpps_person_match_score.utils import standardisation_functions
-from hmpps_person_match_score.utils.db_cleanup import cleanup_splink_tables, generate_view_uuid
 from hmpps_person_match_score.views.base_view import BaseView
 
 
@@ -90,8 +89,8 @@ class MatchView(BaseView):
         row_arrow_1 = pa.Table.from_pandas(row_1, schema=self.SCHEMA, preserve_index=False)
         row_arrow_2 = pa.Table.from_pandas(row_2, schema=self.SCHEMA, preserve_index=False)
 
-        view_uuid_1 = generate_view_uuid()
-        view_uuid_2 = generate_view_uuid()
+        view_uuid_1 = self.generate_view_uuid()
+        view_uuid_2 = self.generate_view_uuid()
 
         linker = DuckDBLinker(
             [row_arrow_1, row_arrow_2],
@@ -108,6 +107,6 @@ class MatchView(BaseView):
         json_output = prediction.as_pandas_dataframe().to_json()
 
         # Clean up Splink tables to avoid OOM, see PR 163
-        cleanup_splink_tables(linker, self.duckdb_connection, [view_uuid_1, view_uuid_2])
+        self.cleanup_splink_tables(linker, [view_uuid_1, view_uuid_2])
 
         return json.loads(json_output)
