+++ hmpps_person_match_score/views/person_match_view.py
@@ -6,7 +6,6 @@ from splink.duckdb.duckdb_linker import DuckDBLinker
 from hmpps_person_match_score.domain.events import Events
 from hmpps_person_match_score.domain.splink_models import SplinkModels
 from hmpps_person_match_score.models.person_match_model import PersonMatching
-from hmpps_person_match_score.utils.db_cleanup import cleanup_splink_tables, generate_view_uuid
 from hmpps_person_match_score.views.base_view import BaseView
 
 
@@ -51,8 +50,8 @@ class PersonMatchView(BaseView):
         dataset_1 = pa.Table.from_pylist(pmm_dict["matching_to"], schema=self.SCHEMA)
         dataset_2 = pa.Table.from_pylist([pmm_dict["matching_from"]], schema=self.SCHEMA)
 
-        view_uuid_1 = generate_view_uuid()
-        view_uuid_2 = generate_view_uuid()
+        view_uuid_1 = self.generate_view_uuid()
+        view_uuid_2 = self.generate_view_uuid()
 
         linker = DuckDBLinker(
             [dataset_1, dataset_2],
@@ -69,6 +68,6 @@ class PersonMatchView(BaseView):
         json_output = prediction.as_pandas_dataframe().to_json()
 
         # Clean up Splink tables to avoid OOM, see PR 163
-        cleanup_splink_tables(linker, self.duckdb_connection, [view_uuid_1, view_uuid_2])
+        self.cleanup_splink_tables(linker, [view_uuid_1, view_uuid_2])
 
         return json.loads(json_output)
